Solved the error not fired click event...
So, I found a mistake. When I click on the ion-item in feed.page.html to open the Publish Page, nothing happens, so I discover that the authguard was called to checkLogin for some user logged, but twice, in the first one returning false, so it redirect again and check again, but now get the user and go to tabs. So in this time the feed page open but the event on the page doesn't works. This commit is a workaround for this problem.